// src/models/employee.ts
export interface Employee {
    accountId: string;
    firstName: string;
    middleName: string;
    lastName: string;
    gender: string;
    state: string;
    city: string;
    wereda: string;
    kebele: string;
    phone: string;
    houseNumber: string;
    subCity: string;
    email: string;
    createdAt?: string;
  }
  